refactor: Define TenantOwnershiptObject with TIdent
