+++ src/meta/app/src/principal/mod.rs
@@ -35,9 +35,12 @@ mod user_setting;
 mod user_stage;
 mod user_stage_file_ident;
 
+mod ownership_object;
+
 pub mod connection_ident;
 pub mod network_policy_ident;
 pub mod password_policy_ident;
+pub mod tenant_ownership_object_ident;
 pub mod user_defined_file_format_ident;
 pub mod user_setting_ident;
 pub mod user_stage_ident;
@@ -47,6 +50,7 @@ pub use file_format::*;
 pub use network_policy::NetworkPolicy;
 pub use network_policy_ident::NetworkPolicyIdent;
 pub use ownership_info::OwnershipInfo;
+pub use ownership_object::OwnershipObject;
 pub use password_policy::PasswordPolicy;
 pub use password_policy_ident::PasswordPolicyIdent;
 pub use principal_identity::PrincipalIdentity;
@@ -54,6 +58,7 @@ pub use role_ident::RoleIdent;
 pub use role_ident::RoleIdentRaw;
 pub use role_info::RoleInfo;
 pub use role_info::RoleInfoSerdeError;
+pub use tenant_ownership_object_ident::TenantOwnershipObjectIdent;
 pub use udf_ident::UdfIdent;
 pub use udf_ident::UdfIdentRaw;
 pub use user_auth::AuthInfo;
@@ -67,8 +72,6 @@ pub use user_defined_function::UDFServer;
 pub use user_defined_function::UserDefinedFunction;
 pub use user_grant::GrantEntry;
 pub use user_grant::GrantObject;
-pub use user_grant::OwnershipObject;
-pub use user_grant::TenantOwnershipObject;
 pub use user_grant::UserGrantSet;
 pub use user_identity::TenantUserIdent;
 pub use user_identity::UserIdentity;
