+++ src/meta/app/src/principal/user_grant.rs
@@ -16,189 +16,10 @@ use std::collections::HashSet;
 use std::fmt;
 use std::ops;
 
-use databend_common_meta_kvapi::kvapi;
 use enumflags2::BitFlags;
 
 use crate::principal::UserPrivilegeSet;
 use crate::principal::UserPrivilegeType;
-use crate::tenant::Tenant;
-
-/// [`OwnershipObject`] is used to maintain the grant object that support rename by id. Using ID over name
-/// have many benefits, it can avoid lost privileges after the object get renamed.
-/// But Stage and UDF do not support the concept of renaming and do not have ids, so names can be used.
-#[derive(serde::Serialize, serde::Deserialize, Clone, Debug, Eq, PartialEq, Hash)]
-pub enum OwnershipObject {
-    /// used on the fuse databases
-    Database {
-        catalog_name: String,
-        db_id: u64,
-    },
-
-    /// used on the fuse tables
-    Table {
-        catalog_name: String,
-        db_id: u64,
-        table_id: u64,
-    },
-
-    Stage {
-        name: String,
-    },
-
-    UDF {
-        name: String,
-    },
-}
-
-impl OwnershipObject {
-    // # Legacy issue: Catalog is not encoded into key.
-    //
-    // But at that time, only `"default"` catalog is used.
-    // Thus we follow the same rule, if catalog is `"default"`, we don't encode it.
-    //
-    // This issue is introduced in https://github.com/drmingdrmer/databend/blob/7681763dc54306e55b5e0326af0510292d244be3/src/query/management/src/role/role_mgr.rs#L86
-    const DEFAULT_CATALOG: &'static str = "default";
-
-    /// Build key with the provided KeyBuilder as sub part of key used to access meta-service
-    pub(crate) fn build_key(&self, b: kvapi::KeyBuilder) -> kvapi::KeyBuilder {
-        match self {
-            OwnershipObject::Database {
-                catalog_name,
-                db_id,
-            } => {
-                // Legacy issue: Catalog is not encoded into key.
-                if catalog_name == Self::DEFAULT_CATALOG {
-                    b.push_raw("database-by-id").push_u64(*db_id)
-                } else {
-                    b.push_raw("database-by-catalog-id")
-                        .push_str(catalog_name)
-                        .push_u64(*db_id)
-                }
-            }
-            OwnershipObject::Table {
-                catalog_name,
-                db_id,
-                table_id,
-            } => {
-                // TODO(flaneur): db_id is not encoded into key. Thus such key can not be decoded.
-                let _ = db_id;
-
-                // Legacy issue: Catalog is not encoded into key.
-                if catalog_name == Self::DEFAULT_CATALOG {
-                    b.push_raw("table-by-id").push_u64(*table_id)
-                } else {
-                    b.push_raw("table-by-catalog-id")
-                        .push_str(catalog_name)
-                        .push_u64(*table_id)
-                }
-            }
-            OwnershipObject::Stage { name } => b.push_raw("stage-by-name").push_str(name),
-            OwnershipObject::UDF { name } => b.push_raw("udf-by-name").push_str(name),
-        }
-    }
-
-    /// Parse encoded key and return the OwnershipObject
-    pub(crate) fn parse_key(p: &mut kvapi::KeyParser) -> Result<Self, kvapi::KeyError> {
-        let q = p.next_raw()?;
-        match q {
-            "database-by-id" => {
-                let db_id = p.next_u64()?;
-                Ok(OwnershipObject::Database {
-                    catalog_name: Self::DEFAULT_CATALOG.to_string(),
-                    db_id,
-                })
-            }
-            "database-by-catalog-id" => {
-                let catalog_name = p.next_str()?;
-                let db_id = p.next_u64()?;
-                Ok(OwnershipObject::Database {
-                    catalog_name,
-                    db_id,
-                })
-            }
-            "table-by-id" => {
-                let table_id = p.next_u64()?;
-                Ok(OwnershipObject::Table {
-                    catalog_name: Self::DEFAULT_CATALOG.to_string(),
-                    db_id: 0,
-                    table_id,
-                })
-            }
-            "table-by-catalog-id" => {
-                let catalog_name = p.next_str()?;
-                let table_id = p.next_u64()?;
-                Ok(OwnershipObject::Table {
-                    catalog_name,
-                    db_id: 0, // string key does not contain db_id
-                    table_id,
-                })
-            }
-            "stage-by-name" => {
-                let name = p.next_str()?;
-                Ok(OwnershipObject::Stage { name })
-            }
-            "udf-by-name" => {
-                let name = p.next_str()?;
-                Ok(OwnershipObject::UDF { name })
-            }
-            _ => Err(kvapi::KeyError::InvalidSegment {
-                i: p.index(),
-                expect: "database-by-id|database-by-catalog-id|table-by-id|table-by-catalog-id|stage-by-name|udf-by-name"
-                    .to_string(),
-                got: q.to_string(),
-            }),
-        }
-        //
-    }
-}
-
-/// The meta-service key of object whose ownership to grant.
-///
-/// It could be a tenant's database, a tenant's table etc.
-/// It is in form of `__fd_object_owners/<tenant>/<object>`.
-/// where `<object>` could be:
-/// - `database-by-id/<db_id>`
-/// - `database-by-catalog-id/<catalog>/<db_id>`
-/// - `table-by-id/<table_id>`
-/// - `table-by-catalog-id/<catalog>/<table_id>`
-/// - `stage-by-name/<stage_name>`
-/// - `udf-by-name/<udf_name>`
-#[derive(Clone, Debug, Eq, PartialEq, Hash)]
-pub struct TenantOwnershipObject {
-    tenant: Tenant,
-    object: OwnershipObject,
-}
-
-impl TenantOwnershipObject {
-    pub fn new(tenant: Tenant, subject: OwnershipObject) -> Self {
-        // Legacy issue: Assert compatibility: No other catalog should be used.
-        // Assertion is disabled.
-        // Instead, accessing field `object` is disallowed.
-        // match &subject {
-        //     OwnershipObject::Database { catalog_name, .. } => {
-        //         assert_eq!(catalog_name, OwnershipObject::DEFAULT_CATALOG);
-        //     }
-        //     OwnershipObject::Table { catalog_name, .. } => {
-        //         assert_eq!(catalog_name, OwnershipObject::DEFAULT_CATALOG);
-        //     }
-        //     OwnershipObject::Stage { .. } => {}
-        //     OwnershipObject::UDF { .. } => {}
-        // }
-
-        Self::new_unchecked(tenant, subject)
-    }
-
-    pub(crate) fn new_unchecked(tenant: Tenant, object: OwnershipObject) -> Self {
-        TenantOwnershipObject { tenant, object }
-    }
-
-    // Because the encoded string key does not contain all of the field:
-    // Catalog and db_id is missing, so we can not rebuild the complete key from string key.
-    // Thus it is not allow to access this field.
-    // pub fn subject(&self) -> &OwnershipObject {
-    //     &self.object
-    // }
-}
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, Eq, PartialEq, Hash)]
 pub enum GrantObject {
@@ -453,208 +274,5 @@ impl fmt::Display for UserGrantSet {
             write!(f, "{}, ", entry)?;
         }
         write!(f, "ROLES: {:?}", self.roles())
-    }
-}
-
-mod kvapi_key_impl {
-    use databend_common_meta_kvapi::kvapi;
-    use databend_common_meta_kvapi::kvapi::KeyBuilder;
-    use databend_common_meta_kvapi::kvapi::KeyError;
-    use databend_common_meta_kvapi::kvapi::KeyParser;
-
-    use crate::principal::user_grant::TenantOwnershipObject;
-    use crate::principal::OwnershipInfo;
-    use crate::principal::OwnershipObject;
-    use crate::tenant::Tenant;
-    use crate::KeyWithTenant;
-
-    impl kvapi::KeyCodec for TenantOwnershipObject {
-        fn encode_key(&self, b: KeyBuilder) -> KeyBuilder {
-            let b = b.push_str(self.tenant_name());
-            self.object.build_key(b)
-        }
-
-        fn decode_key(parser: &mut KeyParser) -> Result<Self, KeyError> {
-            let tenant = parser.next_nonempty()?;
-            let subject = OwnershipObject::parse_key(parser)?;
-
-            Ok(Self {
-                tenant: Tenant::new_nonempty(tenant),
-                object: subject,
-            })
-        }
-    }
-
-    impl kvapi::Key for TenantOwnershipObject {
-        const PREFIX: &'static str = "__fd_object_owners";
-        type ValueType = OwnershipInfo;
-
-        fn parent(&self) -> Option<String> {
-            Some(self.tenant.to_string_key())
-        }
-    }
-
-    impl kvapi::Value for OwnershipInfo {
-        fn dependency_keys(&self) -> impl IntoIterator<Item = String> {
-            []
-        }
-    }
-
-    impl KeyWithTenant for TenantOwnershipObject {
-        fn tenant(&self) -> &Tenant {
-            &self.tenant
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use databend_common_meta_kvapi::kvapi::Key;
-
-    use crate::principal::user_grant::TenantOwnershipObject;
-    use crate::principal::OwnershipObject;
-    use crate::tenant::Tenant;
-    use crate::KeyWithTenant;
-
-    #[test]
-    fn test_tenant_ownership_object_tenant_prefix() {
-        let r =
-            TenantOwnershipObject::new(Tenant::new_literal("tenant"), OwnershipObject::Database {
-                catalog_name: "cat".to_string(),
-                db_id: 1,
-            });
-
-        assert_eq!("__fd_object_owners/tenant/", r.tenant_prefix());
-    }
-
-    #[test]
-    fn test_role_grantee_as_kvapi_key() {
-        // db with default catalog
-        {
-            let role_grantee = TenantOwnershipObject::new_unchecked(
-                Tenant::new_literal("test"),
-                OwnershipObject::Database {
-                    catalog_name: "default".to_string(),
-                    db_id: 1,
-                },
-            );
-
-            let key = role_grantee.to_string_key();
-            assert_eq!("__fd_object_owners/test/database-by-id/1", key);
-
-            let parsed = TenantOwnershipObject::from_str_key(&key).unwrap();
-            assert_eq!(role_grantee, parsed);
-        }
-
-        // db with catalog
-        {
-            let role_grantee = TenantOwnershipObject::new_unchecked(
-                Tenant::new_literal("test"),
-                OwnershipObject::Database {
-                    catalog_name: "cata/foo".to_string(),
-                    db_id: 1,
-                },
-            );
-
-            let key = role_grantee.to_string_key();
-            assert_eq!(
-                "__fd_object_owners/test/database-by-catalog-id/cata%2ffoo/1",
-                key
-            );
-
-            let parsed = TenantOwnershipObject::from_str_key(&key).unwrap();
-            assert_eq!(role_grantee, parsed);
-        }
-
-        // table with default catalog
-        {
-            let obj = TenantOwnershipObject::new_unchecked(
-                Tenant::new_literal("test"),
-                OwnershipObject::Table {
-                    catalog_name: "default".to_string(),
-                    db_id: 1,
-                    table_id: 2,
-                },
-            );
-
-            let key = obj.to_string_key();
-            assert_eq!("__fd_object_owners/test/table-by-id/2", key);
-
-            let parsed = TenantOwnershipObject::from_str_key(&key).unwrap();
-            assert_eq!(
-                TenantOwnershipObject::new_unchecked(
-                    Tenant::new_literal("test"),
-                    OwnershipObject::Table {
-                        catalog_name: "default".to_string(),
-                        db_id: 0, // db_id is not encoded into key
-                        table_id: 2,
-                    }
-                ),
-                parsed
-            );
-        }
-
-        // table with catalog
-        {
-            let obj = TenantOwnershipObject::new_unchecked(
-                Tenant::new_literal("test"),
-                OwnershipObject::Table {
-                    catalog_name: "cata/foo".to_string(),
-                    db_id: 1,
-                    table_id: 2,
-                },
-            );
-
-            let key = obj.to_string_key();
-            assert_eq!(
-                "__fd_object_owners/test/table-by-catalog-id/cata%2ffoo/2",
-                key
-            );
-
-            let parsed = TenantOwnershipObject::from_str_key(&key).unwrap();
-            assert_eq!(
-                TenantOwnershipObject::new_unchecked(
-                    Tenant::new_literal("test"),
-                    OwnershipObject::Table {
-                        catalog_name: "cata/foo".to_string(),
-                        db_id: 0, // db_id is not encoded into key
-                        table_id: 2,
-                    }
-                ),
-                parsed
-            );
-        }
-
-        // stage
-        {
-            let role_grantee = TenantOwnershipObject::new_unchecked(
-                Tenant::new_literal("test"),
-                OwnershipObject::Stage {
-                    name: "foo".to_string(),
-                },
-            );
-
-            let key = role_grantee.to_string_key();
-            assert_eq!("__fd_object_owners/test/stage-by-name/foo", key);
-
-            let parsed = TenantOwnershipObject::from_str_key(&key).unwrap();
-            assert_eq!(role_grantee, parsed);
-        }
-
-        // udf
-        {
-            let role_grantee = TenantOwnershipObject::new_unchecked(
-                Tenant::new_literal("test"),
-                OwnershipObject::UDF {
-                    name: "foo".to_string(),
-                },
-            );
-
-            let key = role_grantee.to_string_key();
-            assert_eq!("__fd_object_owners/test/udf-by-name/foo", key);
-
-            let parsed = TenantOwnershipObject::from_str_key(&key).unwrap();
-            assert_eq!(role_grantee, parsed);
-        }
     }
 }
